共通関数を切り出し
